Refactor respawning out of GameManager and into PlayerRespawner
Also remove unused files and comments
