--- Assets/Scripts/Helpers.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ee87905515e95814dabff3b96ed3a0b2
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
