--- Assets/Scripts/Helpers/MusicHelper.cs
@@ -1,13 +1,0 @@
-
-public class MusicHelper
-{
-    public static void GetSongs()
-    {
-        //Load songs here and send to musicManager.cs
-    }
-
-    public static void GetSoundFX()
-    {
-        //Load soundFX for scene and send to soundFXManager.cs
-    }
-}
