--- Assets/Scripts/Helpers/MusicHelper.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 36727aa0f5807d24b83a57944c3c588f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
