--- Assets/Scripts/Managers/MusicManager.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5dc1329dc4f668b4b87b8cc3eaf5cefa
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
